requirements.txt
fastapi[standard]==0.30.1
langchain
langchain-community
langgraph==0.1.5
langchain_core
langchain-google-genai==1.0.6
google-generativeai==0.5.2
google_genai
python-dotenv==1.0.1
pydantic==2.7.4
dotenv
pycryptodome
-e .
